chore(react): check in progress
