+++ packages/react/src/components/MultiSelect/MultiSelect.stories.js
