+++ packages/react/src/components/MultiSelect/__tests__/FilterableMultiSelect-test.e2e.js
@@ -5,12 +5,15 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import '../../../../index.scss';
+import '../../../../../index.scss';
 
 import React from 'react';
 import { mount } from '@cypress/react';
-import { generateItems, generateGenericItem } from '../../ListBox/test-helpers';
-import FilterableMultiSelect from '../../MultiSelect/FilterableMultiSelect';
+import {
+  generateItems,
+  generateGenericItem,
+} from '../../../ListBox/test-helpers';
+import FilterableMultiSelect from '../FilterableMultiSelect';
 
 describe('FilterableMultiSelect', () => {
   beforeEach(() => {
