+++ packages/react/src/components/MultiSelect/__tests__/FilterableMultiSelect-test.js
@@ -7,14 +7,14 @@
 
 import React from 'react';
 import { mount } from 'enzyme';
-import FilterableMultiSelect from '../../MultiSelect/FilterableMultiSelect';
+import FilterableMultiSelect from '../FilterableMultiSelect';
 import {
   assertMenuOpen,
   assertMenuClosed,
   findMenuIconNode,
   generateItems,
   generateGenericItem,
-} from '../../ListBox/test-helpers';
+} from '../../../ListBox/test-helpers';
 
 const listItemName = 'ListBoxMenuItem';
 const openMenu = (wrapper) => {
@@ -25,8 +25,6 @@ describe('FilterableMultiSelect', () => {
   let mockProps;
 
   beforeEach(() => {
-    // jest.mock('../../../internal/deprecateFieldOnObject');
-
     mockProps = {
       id: 'test-filterable-multiselect',
       disabled: false,
